pyproject.toml
[project]
name = "a2a-langgraph-demo"
version = "0.1.0"
description = "LangGraph currency agent with A2A Protocol and MySQL storage"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "a2a-sdk[mysql]==0.3.0",
    "click>=8.1.8",
    "httpx>=0.28.1",
    "langchain-google-genai>=2.0.10",
    "langgraph>=0.3.18",
    "langchain-openai>=0.1.0",
    "pydantic>=2.10.6",
    "python-dotenv>=1.1.0",
    "uvicorn>=0.34.2",
    "sse-starlette>=2.3.6",
    "starlette>=0.46.2",
    "sqlalchemy>=2.0.0",
    "pymysql>=1.1.0",
    "cryptography>=42.0.0",  # Required for PyMySQL
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"